Add Last Updated field to interactive table
Most plugins are hosted on GitHub so for now we're using the date of
the last commit on the default branch as a proxy for a last updated
time. Plugins not on GitHub get no love at the moment.

In order to produce the "10 days ago" text a new dependency on package
"moment.js" has been added. MIT license, see DEPENDENCIES.md for info.